compilation error fix

+++ src/ee/expressions/subqueryexpression.cpp
@@ -57,8 +57,8 @@ namespace voltdb {
         VOLT_TRACE ("Running subquery: %d", m_subqueryId);
         // Substitute parameters
         if (m_tveParams.get() != NULL) {
-            int paramsCnt = m_tveParams->size();
-            for (int i = 0; i < paramsCnt; ++i) {
+            size_t paramsCnt = m_tveParams->size();
+            for (size_t i = 0; i < paramsCnt; ++i) {
                 AbstractExpression* tveParam = (*m_tveParams)[i];
                 NValue param = tveParam->eval(tuple1, tuple2);
                 (*m_parameterContainer)[m_paramIdxs[i]] = param;
